ROL: correct warnings in examples and tests
The tests were mostly signed/unsigned comparisons
which I cast to match.  I did consider changing
the types, but other comparisons in the same code
would have required a cast in the opposite direction
so that was pointless.  The example code was
complaining about unused variables. I did lleave those
in place but commented out so that the example is still
there but the compiler does not complain.  If someone
wanted to instead use those in some way that would be good.

closes issue #2284

+++ packages/rol/src/vector/ROL_PinTVector.hpp
@@ -193,7 +193,7 @@ protected:
     int numLeft = 0;
     int numRight = 0;
  
-    for(int i=0;i<stencil_.size();i++) {
+    for(int i=0;i<(int)stencil_.size();i++) {
       if(stencil_[i]<0) 
         numLeft = std::max(numLeft,std::abs(stencil_[i]));
       else if(stencil_[i]>0) 
@@ -288,11 +288,11 @@ public:
       return true;
 
     // these are "neighbor" unowned vectors (left)
-    if(-leftVectors_.size() <= i && i < 0)
+    if(-(int)leftVectors_.size() <= i && i < 0)
       return true;
 
     // these are "neighbor" unowned vectors (right)
-    if(numOwnedSteps() <= i && i<numOwnedSteps()+rightVectors_.size())
+    if(numOwnedSteps() <= i && i<(int)(numOwnedSteps()+rightVectors_.size()))
       return true;
 
     return false;
@@ -309,11 +309,11 @@ public:
       return stepVectors_->get(i);
 
     // these are "neighbor" unowned vectors (left)
-    if(-leftVectors_.size() <= i && i < 0) 
+    if(-(int)leftVectors_.size() <= i && i < 0) 
       return leftVectors_[leftVectors_.size()+i];
 
     // these are "neighbor" unowned vectors (right)
-    if(numOwnedSteps() <= i && i<numOwnedSteps()+rightVectors_.size())
+    if((int)numOwnedSteps() <= i && i<(int)(numOwnedSteps()+rightVectors_.size()))
       return rightVectors_[i-numOwnedSteps()];
 
     return nullPtr;
